Image (#12)
* Image and pixel transformation

* Pixel transformation

* WtfImageBuilder

* Use the correct H value

* Code inspection

+++ api/src/main/java/wtf/file/api/color/value/DefaultValue.java
@@ -15,6 +15,6 @@ public interface DefaultValue {
      * @param channelWidth the bit width of the color channel
      * @return the default value as a {@link Long}
      */
-    Long forChannelWidth(short channelWidth);
+    short forChannelWidth(int channelWidth);
 
 }
